This is a final project part 2 - best personal solution
And a branch set for block 3
